refactor: mv all helper-fns to src/utils/ dir
