+++ src/sanitize/alphanumericWithSpaces.ts
@@ -1,4 +1,4 @@
-import { getSanitizerFn } from "./_helpers.js";
+import { getSanitizerFn } from "../utils/getSanitizerFn.js";
 
 /**
  * Removes all non-alphanumeric/space characters from `str`.
