+++ src/sanitize/handle.ts
@@ -1,4 +1,4 @@
-import { getSanitizerFn } from "./_helpers.js";
+import { getSanitizerFn } from "../utils/getSanitizerFn.js";
 
 /**
  * Remove all non-handle characters from `str` (e.g., "@foo_user").
