+++ src/sanitize/hexadecimal.ts
@@ -1,4 +1,4 @@
-import { getSanitizerFn } from "./_helpers.js";
+import { getSanitizerFn } from "../utils/getSanitizerFn.js";
 
 /**
  * Remove all non-hexadecimal characters from `str`.
