+++ src/sanitize/index.ts
@@ -1,4 +1,3 @@
-export { getSanitizerFn } from "./_helpers.js";
 export * from "./alphabetic.js";
 export * from "./alphabeticWithSpaces.js";
 export * from "./alphanumeric.js";
