+++ src/sanitize/jwt.ts
@@ -1,4 +1,4 @@
-import { getSanitizerFn } from "./_helpers.js";
+import { getSanitizerFn } from "../utils/getSanitizerFn.js";
 
 /**
  * Remove characters from `str` which are not valid in a JSON Web Token.
