+++ src/sanitize/phone.test.ts
@@ -2,7 +2,7 @@ import {
   ZERO_WIDTH_SPACE,
   NULL_CONTROL_CHAR,
   ZERO_WIDTH_AND_CONTROL_CHAR_REGEX,
-} from "./_helpers.js";
+} from "../utils/characters.test.js";
 import { sanitizePhone } from "./phone.js";
 
 describe("sanitize: phone", () => {
