+++ src/validate/index.ts
@@ -1,4 +1,3 @@
-export { getValidatorFn } from "./_helpers.js";
 export * from "./alphabetic.js";
 export * from "./alphabeticWithSpaces.js";
 export * from "./alphanumeric.js";
